Separation of components to their files
Separation of components to their own relative files in components folder
